+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+### v3.5.0
+* added `fixedHeight` option, allows you to have equal weeks number in every month
+* added method `disableDate`, allows you to disabled one or multiple dates with datepicker API
+* added prop `disabledDates` - it is a Set which holds all disabled dates
+* added possibility to pass `{silent: true}` to `update` and `setCurrentView` methods, [#583](https://github.com/t1m0n/air-datepicker/issues/568)
+* changed `update` method - now if you pass `selectedDates` here then calendar will keep selected only those dates
+* fixed selecting time on same date when `range: true`, [#568](https://github.com/t1m0n/air-datepicker/issues/568)
+* fixed date conversion to local date when using strings, e.g `selectDate('2024-03-05')` [#589](https://github.com/t1m0n/air-datepicker/issues/589)
+
 ### v3.4.0
 * added new options `onFocus` and `onBeforeSelect` grant you more control over range selection behaviour and more [#526](https://github.com/t1m0n/air-datepicker/issues/526)
 * added new method `getViewDates()` allows you to get all dates that should be currently displayed in calendar [#536](https://github.com/t1m0n/air-datepicker/issues/536)